qemu-plugin: implement the atexit_cb() functionality
Based on the existing code for other callback handlers, add the
_atexit_cb() handling, so that plugins can implement on_exit()
from the HasCallbacks trait. The on_exit() function is called
once on unloading of the plugin.

The at exit callback exists in v1 of the plugin specification,
so it has not got any "feature = plugin-api-vX" gaurds at all.

Tested using qemu-user, untested on qemu-system.

Signed-off-by: Harry van Haaren <harry.vanhaaren@openchip.com>